Contacts Fragment added

+++ app/src/main/java/io/github/thefrsh/stratus/transfer/UserTransfer.kt
@@ -1,3 +1,6 @@
 package io.github.thefrsh.stratus.transfer
 
-class UserTransfer(val id: Int, val username: String)
+class UserTransfer(
+    val id: Long,
+    val username: String
+)
